+++ src/main/kotlin/kr/hhplus/be/server/application/ProductApplication.kt
@@ -1,15 +1,14 @@
 package kr.hhplus.be.server.application
 
-import com.fasterxml.jackson.databind.ObjectMapper
 import kr.hhplus.be.server.application.model.PopularProductInfo
 import kr.hhplus.be.server.common.utils.LocalDateTimeTruncator
 import kr.hhplus.be.server.domain.order.OrderProductService
+import kr.hhplus.be.server.domain.product.PopularProductService
 import kr.hhplus.be.server.domain.product.ProductService
 import kr.hhplus.be.server.domain.product.model.Product
 import kr.hhplus.be.server.domain.product.model.ProductStatus
 import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.data.domain.Pageable
-import org.springframework.data.redis.core.RedisTemplate
 import org.springframework.stereotype.Service
 import java.time.LocalDateTime
 
@@ -17,8 +16,7 @@ import java.time.LocalDateTime
 class ProductApplication @Autowired constructor (
     val productService: ProductService,
     val orderProductService: OrderProductService,
-    val redisTemplate: RedisTemplate<String, String>,
-    val objectMapper: ObjectMapper
+    val popularProductService: PopularProductService,
 ) {
 
     fun getProducts(page: Pageable, status: ProductStatus?) : List<Product> {
@@ -43,25 +41,12 @@ class ProductApplication @Autowired constructor (
 
     fun getPopularProducts(): List<PopularProductInfo> {
         return runCatching {
-            getPopularProductsFromCache()
+            popularProductService.getPopularProductsFromCache()
         }.getOrElse {
             getPopularProductsFromDatabase()
         }.also {
-            setPopularProductsToCache(it)
+            popularProductService.setPopularProductsToCache(it)
         }
     }
 
-    fun getPopularProductsFromCache() : List<PopularProductInfo> {
-        val operation = redisTemplate.opsForValue()
-        val json: String = operation["popularProducts"] ?: throw NoSuchElementException()
-        val data: List<PopularProductInfo> = objectMapper.readValue(json, Array<PopularProductInfo>::class.java).toList()
-
-        return data
-    }
-
-    fun setPopularProductsToCache(popularProducts: List<PopularProductInfo>) {
-        val operation = redisTemplate.opsForValue()
-        val json = objectMapper.writeValueAsString(popularProducts)
-        operation["popularProducts"] = json
-    }
 }
